requirements.txt
selenium>=4.16.0
webdriver-manager>=4.0.1
python-dotenv>=1.0.0
google-generativeai>=0.3.0
PyPDF2>=3.0.1
beautifulsoup4>=4.12.0
pandas>=2.1.0
pytest>=7.4.0
pytest-mock>=3.12.0
